At the moment, the app have a version without progress bar; which works well. and another version with a new view displayed at the start but this one isn't working due to two problem: The LoadingViewController isn't handling my download state and I don't know how to show the Master/Detail from this one.
In this commit, I added another loader, that one is temporary while I fix the two bugs said upper.

I also start to implement the progress bar
The loader should work but I need to fix the two bugs to check it as well

PS: Check comments in the viewDidLoad method of the MasterViewController in order to make the app working.